Split DSMatchType into its own source file
