+++ lib/go-tc/enum/enum.go
@@ -220,46 +220,3 @@ func (e QStringIgnore) String() string {
 		return "INVALID"
 	}
 }
-
-type DSMatchType string
-
-const (
-	DSMatchTypeHostRegex     DSMatchType = "HOST_REGEXP"
-	DSMatchTypePathRegex     DSMatchType = "PATH_REGEXP"
-	DSMatchTypeSteeringRegex DSMatchType = "STEERING_REGEXP"
-	DSMatchTypeHeaderRegex   DSMatchType = "HEADER_REGEXP"
-	DSMatchTypeInvalid       DSMatchType = ""
-)
-
-// String returns a string representation of this delivery service match type.
-func (t DSMatchType) String() string {
-	switch t {
-	case DSMatchTypeHostRegex:
-		fallthrough
-	case DSMatchTypePathRegex:
-		fallthrough
-	case DSMatchTypeSteeringRegex:
-		fallthrough
-	case DSMatchTypeHeaderRegex:
-		return string(t)
-	default:
-		return "INVALID_MATCH_TYPE"
-	}
-}
-
-// DSMatchTypeFromString returns a delivery service match type object from its string representation, or DSMatchTypeInvalid if the string is not a valid type.
-func DSMatchTypeFromString(s string) DSMatchType {
-	s = strings.ToLower(strings.Replace(s, "_", "", -1))
-	switch s {
-	case "hostregexp":
-		return DSMatchTypeHostRegex
-	case "pathregexp":
-		return DSMatchTypePathRegex
-	case "steeringregexp":
-		return DSMatchTypeSteeringRegex
-	case "headerregexp":
-		return DSMatchTypeHeaderRegex
-	default:
-		return DSMatchTypeInvalid
-	}
-}
